servicio_expedientes/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<3.16
requests